Fix `FromIterable` so it supports sending tuples
When sending tuples as input to the `AsyncObservable.from_iterable` method
a TypeError exception was raised. The exception was caused by the following line:
```python
log.debug("sync_worker. asending: %s" % value)
```

That is solved by either passing value as a tuple with one value:
```python
log.debug("sync_worker. asending: %s" % (value,))
```

Or by using the parameters in the debug method, I choose the second option
as it is what I know to be the best practice when using log methods.